+++ libraries/exoplayer/src/test/java/androidx/media3/exoplayer/DefaultLoadControlTest.java
@@ -18,6 +18,7 @@ package androidx.media3.exoplayer;
 import static com.google.common.truth.Truth.assertThat;
 
 import androidx.media3.common.C;
+import androidx.media3.common.Timeline;
 import androidx.media3.common.util.Util;
 import androidx.media3.exoplayer.DefaultLoadControl.Builder;
 import androidx.media3.exoplayer.source.TrackGroupArray;
@@ -179,7 +180,12 @@ public class DefaultLoadControlTest {
   @Test
   public void shouldContinueLoading_withNoSelectedTracks_returnsTrue() {
     loadControl = builder.build();
-    loadControl.onTracksSelected(new Renderer[0], TrackGroupArray.EMPTY, new ExoTrackSelection[0]);
+    loadControl.onTracksSelected(
+        Timeline.EMPTY,
+        LoadControl.EMPTY_MEDIA_PERIOD_ID,
+        new Renderer[0],
+        TrackGroupArray.EMPTY,
+        new ExoTrackSelection[0]);
 
     assertThat(
             loadControl.shouldContinueLoading(
@@ -203,6 +209,8 @@ public class DefaultLoadControlTest {
 
     assertThat(
             loadControl.shouldStartPlayback(
+                Timeline.EMPTY,
+                LoadControl.EMPTY_MEDIA_PERIOD_ID,
                 MIN_BUFFER_US,
                 SPEED,
                 /* rebuffering= */ false,
@@ -222,6 +230,8 @@ public class DefaultLoadControlTest {
 
     assertThat(
             loadControl.shouldStartPlayback(
+                Timeline.EMPTY,
+                LoadControl.EMPTY_MEDIA_PERIOD_ID,
                 /* bufferedDurationUs= */ 2_999_999,
                 SPEED,
                 /* rebuffering= */ false,
@@ -229,6 +239,8 @@ public class DefaultLoadControlTest {
         .isFalse();
     assertThat(
             loadControl.shouldStartPlayback(
+                Timeline.EMPTY,
+                LoadControl.EMPTY_MEDIA_PERIOD_ID,
                 /* bufferedDurationUs= */ 3_000_000,
                 SPEED,
                 /* rebuffering= */ false,
@@ -247,6 +259,8 @@ public class DefaultLoadControlTest {
 
     assertThat(
             loadControl.shouldStartPlayback(
+                Timeline.EMPTY,
+                LoadControl.EMPTY_MEDIA_PERIOD_ID,
                 /* bufferedDurationUs= */ 499_999,
                 SPEED,
                 /* rebuffering= */ true,
@@ -254,6 +268,8 @@ public class DefaultLoadControlTest {
         .isFalse();
     assertThat(
             loadControl.shouldStartPlayback(
+                Timeline.EMPTY,
+                LoadControl.EMPTY_MEDIA_PERIOD_ID,
                 /* bufferedDurationUs= */ 500_000,
                 SPEED,
                 /* rebuffering= */ true,
@@ -273,6 +289,8 @@ public class DefaultLoadControlTest {
 
     assertThat(
             loadControl.shouldStartPlayback(
+                Timeline.EMPTY,
+                LoadControl.EMPTY_MEDIA_PERIOD_ID,
                 /* bufferedDurationUs= */ 3_999_999,
                 SPEED,
                 /* rebuffering= */ true,
@@ -280,6 +298,8 @@ public class DefaultLoadControlTest {
         .isFalse();
     assertThat(
             loadControl.shouldStartPlayback(
+                Timeline.EMPTY,
+                LoadControl.EMPTY_MEDIA_PERIOD_ID,
                 /* bufferedDurationUs= */ 4_000_000,
                 SPEED,
                 /* rebuffering= */ true,
@@ -298,6 +318,8 @@ public class DefaultLoadControlTest {
 
     assertThat(
             loadControl.shouldStartPlayback(
+                Timeline.EMPTY,
+                LoadControl.EMPTY_MEDIA_PERIOD_ID,
                 /* bufferedDurationUs= */ 499_999,
                 SPEED,
                 /* rebuffering= */ true,
@@ -305,6 +327,8 @@ public class DefaultLoadControlTest {
         .isFalse();
     assertThat(
             loadControl.shouldStartPlayback(
+                Timeline.EMPTY,
+                LoadControl.EMPTY_MEDIA_PERIOD_ID,
                 /* bufferedDurationUs= */ 500_000,
                 SPEED,
                 /* rebuffering= */ true,
@@ -315,7 +339,8 @@ public class DefaultLoadControlTest {
   private void build() {
     builder.setAllocator(allocator).setTargetBufferBytes(TARGET_BUFFER_BYTES);
     loadControl = builder.build();
-    loadControl.onTracksSelected(new Renderer[0], null, null);
+    loadControl.onTracksSelected(
+        Timeline.EMPTY, LoadControl.EMPTY_MEDIA_PERIOD_ID, new Renderer[0], null, null);
   }
 
   private void makeSureTargetBufferBytesReached() {